fix conglicts whth const and non-const iterators, typenames
